requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
pyyaml==6.0.1
rich==13.7.0
pandas==2.1.3
python-dotenv==1.0.0
requests==2.31.0